+++ ase/calculators/amber.py
@@ -30,8 +30,10 @@ class Amber(FileIOCalculator):
     """
 
     implemented_properties = ['energy', 'forces']
+    discard_results_on_any_change = True
 
-    def __init__(self, restart=None, ignore_bad_restart_file=False,
+    def __init__(self, restart=None,
+                 ignore_bad_restart_file=FileIOCalculator._deprecated,
                  label='amber', atoms=None, command=None,
                  amber_exe='sander -O ',
                  infile='mm.in', outfile='mm.out',
@@ -97,11 +99,6 @@ class Amber(FileIOCalculator):
 
         FileIOCalculator.__init__(self, restart, ignore_bad_restart_file,
                                   label, atoms, **kwargs)
-
-    def set(self, **kwargs):
-        changed_parameters = FileIOCalculator.set(self, **kwargs)
-        if changed_parameters:
-            self.reset()
 
     def write_input(self, atoms=None, properties=None, system_changes=None):
         """Write updated coordinates to a file."""