+++ ase/calculators/mopac.py
@@ -19,6 +19,7 @@ from ase.units import kcal, mol, Debye
 class MOPAC(FileIOCalculator):
     implemented_properties = ['energy', 'forces', 'dipole', 'magmom']
     command = 'mopac PREFIX.mop 2> /dev/null'
+    discard_results_on_any_change = True
 
     default_parameters = dict(
         method='PM7',
@@ -30,7 +31,8 @@ class MOPAC(FileIOCalculator):
                'PM6-DH2', 'PM6-DH2X', 'PM6-D3H4', 'PM6-D3H4X', 'PMEP', 'PM7',
                'PM7-TS', 'RM1']
 
-    def __init__(self, restart=None, ignore_bad_restart_file=False,
+    def __init__(self, restart=None,
+                 ignore_bad_restart_file=FileIOCalculator._deprecated,
                  label='mopac', atoms=None, **kwargs):
         """Construct MOPAC-calculator object.
 
@@ -67,11 +69,6 @@ class MOPAC(FileIOCalculator):
         """
         FileIOCalculator.__init__(self, restart, ignore_bad_restart_file,
                                   label, atoms, **kwargs)
-
-    def set(self, **kwargs):
-        changed_parameters = FileIOCalculator.set(self, **kwargs)
-        if changed_parameters:
-            self.reset()
 
     def write_input(self, atoms, properties=None, system_changes=None):
         FileIOCalculator.write_input(self, atoms, properties, system_changes)