+++ ase/gui/gui.py
@@ -1,8 +1,6 @@
-import os
 import pickle
 import subprocess
 import sys
-import tempfile
 import weakref
 from functools import partial
 from ase.gui.i18n import _
@@ -400,13 +398,8 @@ class GUI(View, Status):
         return save_dialog(self)
 
     def external_viewer(self, name):
-        command = {'xmakemol': 'xmakemol -f',
-                   'rasmol': 'rasmol -xyz'}.get(name, name)
-        fd, filename = tempfile.mkstemp('.xyz', 'ase.gui-')
-        os.close(fd)
-        self.images.write(filename)
-        os.system('(%s %s &); (sleep 60; rm %s) &' %
-                  (command, filename, filename))
+        from ase.visualize import view
+        return view(list(self.images), viewer=name)
 
     def selected_atoms(self):
         selection_mask = self.images.selected[:len(self.atoms)]