+++ ase/test/fio/test_oi.py
@@ -103,18 +103,18 @@ def test_ioformat(format, atoms, catch_warnings):
 
     io = ioformats[format]
     print('{0:20}{1}{2}{3}{4}'.format(format,
-                                      ' R'[bool(io.read)],
-                                      ' W'[bool(io.write)],
+                                      ' R'[io.can_read],
+                                      ' W'[io.can_write],
                                       '+1'[io.single],
                                       'SF'[io.acceptsfd]))
     fname1 = 'io-test.1.{}'.format(format)
     fname2 = 'io-test.2.{}'.format(format)
-    if io.write:
+    if io.can_write:
         write(fname1, atoms, format=format)
         if not io.single:
             write(fname2, images, format=format)
 
-        if io.read:
+        if io.can_read:
             for a in [read(fname1, format=format), read(fname1)]:
                 check(a, atoms, format)