+++ ase/test/md/test_md_logger.py
@@ -73,7 +73,9 @@ def test_md(cls_and_kwargs, atoms=dimer, calc=calc, logfile="md.log",
     md.run(steps=5)
 
     # assert log file has correct length
-    length = sum(1 for l in open(logfile))
+    with open(logfile) as fd:
+        length = len(fd.readlines())
+
     assert length == 12, length
 
     with Trajectory(trajectory) as traj, open(logfile) as f: